+++ VASTvaporizer/Source/Engine/FX/VASTFormantFilter.cpp
@@ -119,7 +119,7 @@ void CVASTFormantFilter::releaseResources() {
 
 void CVASTFormantFilter::parameterChanged(const String& parameterID, float newValue) {
 	if (parameterID.startsWith("m_bFormantOnOff")) {
-		if (newValue == SWITCH::SWITCH_ON)
+		if (newValue == static_cast<int>(SWITCH::SWITCH_ON))
 			switchOn();
 		else
 			switchOff();
