+++ VASTvaporizer/Source/Engine/FX/VASTWaveshaper.cpp
@@ -119,7 +119,7 @@ CVASTWaveshaper::~CVASTWaveshaper(void) {
 
 void CVASTWaveshaper::parameterChanged(const String& parameterID, float newValue) {
 	if (parameterID.startsWith("m_bWaveshaperOnOff")) {
-		if (newValue == SWITCH::SWITCH_ON)
+		if (newValue == static_cast<int>(SWITCH::SWITCH_ON))
 			switchOn();
 		else
 			switchOff();
